added dynamic scaling
