+++ frontend/src/App.css
@@ -30,55 +30,14 @@ body {
     background: inherit; /* Make sure the gradient background is inherited */
 }
 
-.navbar {
-    background-color: #333;
-    width: 200px;
-    height: 100%;
-    position: fixed;
-    top: 0;
-    left: 0;
-    overflow: auto; /* Ensure the navbar scrolls if content overflows */
-}
-
-.navbar ul {
-    list-style-type: none;
-    margin: 0;
-    padding: 0;
-}
-
-.navbar li {
-    width: 100%;
-}
-
-.navbar li a {
-    display: block;
-    color: white;
-    text-align: left;
-    padding: 14px 16px;
-    text-decoration: none;
-}
-
-.navbar li a:hover {
-    background-color: #282828;
-}
-
 .main-content {
-    margin-left: 200px;
-    padding: 20px;
-    flex-grow: 1;
     display: flex;
     flex-direction: column;
     justify-content: flex-start;
     align-items: flex-start;
-    background: inherit; /* Make sure the gradient background is inherited */
+    background: inherit; /* Ensure background is inherited */
     overflow: auto;
-}
-
-.logo {
-    height: 6em;
-    padding: 1.5em;
-    will-change: filter;
-    transition: filter 300ms;
+    box-sizing: border-box;
 }
 
 .logo:hover {
@@ -102,10 +61,6 @@ body {
     a:nth-of-type(2) .logo {
         animation: logo-spin infinite 20s linear;
     }
-}
-
-.card {
-    padding: 2em;
 }
 
 .read-the-docs {
