+++ frontend/src/Main.css
@@ -30,7 +30,6 @@ body {
     margin: 0;
     display: flex;
     flex-direction: column; /* Ensure flex direction is column for body */
-    min-width: 320px;
     min-height: 100vh;
     background: inherit; /* Ensure body inherits the background */
 }
@@ -52,7 +51,6 @@ body {
 
 .navbar {
     background-color: #333;
-    width: 200px;
     height: 100vh;
     position: fixed;
     top: 0;
@@ -73,7 +71,6 @@ body {
     display: block;
     color: white;
     text-align: left;
-    padding: 14px 16px;
     text-decoration: none;
 }
 
@@ -82,8 +79,6 @@ body {
 }
 
 .main-content {
-    margin-left: 200px;
-    padding: 20px;
     flex-grow: 1;
     display: flex;
     justify-content: center;
@@ -94,7 +89,6 @@ body {
 
 .logo {
     height: 6em;
-    padding: 1.5em;
     will-change: filter;
     transition: filter 300ms;
 }
