+++ frontend/src/components/Layout.jsx
@@ -1,7 +1,6 @@
 // src/components/Layout.jsx
 import React from "react";
 import Navbar from "./Navbar";
-import ProfileIcon from "./ProfileIcon";
 import { Outlet } from "react-router-dom";
 
 /**
@@ -11,11 +10,8 @@ const Layout = () => {
     return (
         <div>
             <Navbar />
-            <ProfileIcon />
             <div className="main-content">
-                <div id="root">
-                    <Outlet />
-                </div>
+                <Outlet />
             </div>
         </div>
     );
