+++ frontend/src/components/Navbar.css
@@ -16,7 +16,7 @@ body {
 
 body .navbar {
     background: linear-gradient(to bottom, #242424, #000000) !important;
-    width: 270px;
+    width: 15vw;
     height: 100vh;
     position: fixed;
     top: 0;
@@ -25,23 +25,20 @@ body .navbar {
 
 .navbar {
     background: linear-gradient(to bottom, #242424, #000000);
-    width: 270px;
     height: 100vh;
-    position: fixed;
     top: 0;
     left: 0;
     display: flex;
     flex-direction: column;
     align-items: center;
-    padding-top: 20px;
     box-sizing: border-box;
 }
 
 .navbar-title {
-    font-size: 2.75rem;
+    font-size: 3vw;
     font-weight: bold;
     color: white;
-    margin-top: 50px;
+    margin-top: 10vh;
     text-align: center;
 }
 
@@ -51,7 +48,7 @@ body .navbar {
 }
 
 .navbar-logo img {
-    width: 150px;
+    width: 11vw;
 }
 
 .navbar ul {
@@ -70,9 +67,10 @@ body .navbar {
 .navbar li a {
     display: block;
     color: white;
-    text-align: center;
-    padding: 14px 16px;
+    text-align: left;
+    padding: 1vh 1vh;
     text-decoration: none;
+    font-size: 1.2vw;
 }
 
 .navbar li a:hover,
@@ -81,8 +79,6 @@ body .navbar {
 }
 
 .main-content {
-    margin-left: 270px;
-    padding: 20px;
     flex-grow: 1;
     height: 100vh;
     overflow: auto;
@@ -90,7 +86,6 @@ body .navbar {
 }
 
 #root {
-    max-width: 1280px;
     width: 100%;
     margin: 0 auto;
     text-align: center;
@@ -98,7 +93,6 @@ body .navbar {
 
 .logo {
     height: 6em;
-    padding: 1.5em;
     will-change: filter;
     transition: filter 300ms;
 }
@@ -114,32 +108,8 @@ body .navbar {
 .navbar-settings {
     width: 100%;
     text-align: center;
-    margin-bottom: 30px;
 }
 
 .navbar-settings img {
-    width: 40px;
-}
-
-@keyframes logo-spin {
-    from {
-        transform: rotate(0deg);
-    }
-    to {
-        transform: rotate(360deg);
-    }
-}
-
-@media (prefers-reduced-motion: no-preference) {
-    a:nth-of-type(2) .logo {
-        animation: logo-spin infinite 20s linear;
-    }
-}
-
-.card {
-    padding: 2em;
-}
-
-.read-the-docs {
-    color: #888;
+    width: 3vw;
 }
