--- frontend/src/components/ProfileIcon.css
@@ -1,18 +1,0 @@
-/* src/components/ProfileIcon.css */
-.profile-icon {
-    position: fixed;
-    top: 30px;
-    right: 30px;
-    width: 60px;
-    height: 60px;
-    border-radius: 50%;
-    overflow: hidden;
-    cursor: pointer;
-    border: 2px solid white;
-}
-
-.profile-icon img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
