--- frontend/src/pages/Profile.css
@@ -1,175 +1,0 @@
-.profile-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 20px;
-}
-
-.profile-title {
-    width: 100%;
-    margin-bottom: 20px;
-    margin-top: 510px;
-    text-align: left;
-}
-
-.profile-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-start;
-    border-bottom: 2px solid #ccc;
-    padding-bottom: 20px;
-    width: 100%;
-    margin-bottom: 20px;
-}
-
-.profile-details-left {
-    display: flex;
-    flex-direction: column;
-}
-
-.profile-info {
-    display: flex;
-    align-items: center;
-}
-
-.profile-pic img {
-    border-radius: 50%;
-    margin-right: 50px; /* Increased margin for more space */
-}
-
-.profile-details {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-}
-
-.name-username {
-    display: flex;
-    align-items: baseline;
-}
-
-.profile-name {
-    font-size: 24px; /* Increased font size for the profile name */
-    font-weight: bold; /* Bold the profile name */
-    margin-right: 10px; /* Space between name and username */
-}
-
-.profile-username {
-    font-size: 16px; /* Adjusted font size for the username */
-    color: gray; /* Optional: change color for distinction */
-}
-
-.profile-counters {
-    font-size: 16px; /* Adjusted font size for the counters */
-    margin-top: 10px; /* Space between username and counters */
-}
-
-.bio-section {
-    margin-top: 20px;
-}
-
-.edit-bio-section {
-    display: flex;
-    align-items: center;
-}
-
-.bio-textarea {
-    width: 300px;
-    height: 100px;
-    padding: 10px;
-    margin-right: 10px;
-    border-radius: 5px;
-    border: 1px solid #ccc;
-    font-size: 14px;
-}
-
-.save-bio-button {
-    padding: 10px 20px;
-    background-color: #f0f0f0;
-    color: black;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    cursor: pointer;
-}
-
-.save-bio-button:hover {
-    background-color: #e0e0e0;
-}
-
-.edit-bio-button {
-    align-self: flex-start;
-}
-
-.edit-bio-button button {
-    padding: 10px 20px;
-    background-color: #f0f0f0;
-    color: black;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    cursor: pointer;
-}
-
-.edit-bio-button button:hover {
-    background-color: #e0e0e0;
-}
-
-.content-row {
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-}
-
-.playlist,
-.achievement {
-    margin: 10px;
-    text-align: center; /* Center the text for playlist titles */
-}
-
-.playlist img {
-    border-radius: 10px;
-    display: block;
-    margin: 0 auto; /* Center the image */
-}
-
-.playlist p,
-.achievement p {
-    margin-top: 5px;
-    font-size: 14px;
-}
-
-.playlist-link {
-    color: white;
-    text-decoration: underline;
-}
-
-.playlist-link:hover {
-    color: #888888;
-}
-
-.achievements {
-    padding: 0 20px; /* Add padding to ensure achievements don't get cut off */
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: left; /* Align items to the start */
-    gap: 10px; /* Space between achievements */
-}
-
-.achievement {
-    background-color: #333;
-    color: white;
-    padding: 20px;
-    border-radius: 10px;
-    width: 18%;
-    text-align: center; /* Center text */
-    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1); /* Optional: add shadow for depth */
-    margin: 10px 0; /* Space above and below each achievement */
-}
-
-.achievement-title {
-    font-weight: bold;
-    margin-bottom: 5px;
-}
-
-.achievement-description {
-    font-size: 14px;
-}
