+++ frontend/src/utils/Spotify.js
@@ -1,0 +1,32 @@
+import axios from "axios";
+
+/**
+ * The Spotify class is responsible for pulling data from the
+ * Spotify Web API using client credentials
+ */
+class Spotify {
+    static accessToken = "";
+
+    static refreshAccess() {
+        body = {
+            method: "post",
+            
+        }
+    }
+
+    static makeRequest() {
+
+    }
+
+    static refreshPlaylistCovers(playlists) {
+
+    }
+
+    static refreshTrackCovers(playlist) {
+
+    }
+
+    static getTrendingTracks() [
+
+    ]
+}
